Removing fallacious todo item
There's no plausible use case for dynamically mixing in 'BadCoordinate' (due to the fact that 'x', 'y', etc. still need to be specified on the fly).  There _is_ a use case for doing it with 'PriorityCoordinate', but it does not seem worth it to me to replace the existing 8 lines of code with an incoherent, untamable, experimental 30-line blob of macro-using code.